Add the possibility to use the index on complex values.
Now it is possible to pass a function to new/1 and new/2, the function will be
executed on every value on the SortedSet, the output of the function will be
the index, while the input will be the value.

The function is set as identity function, so it will work as expeceted in
simple case.